app-zaliczenie/requirements.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.2
packaging==24.0
pycparser==2.22
PyJWT==2.8.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
Werkzeug==3.0.3
